move caching into SW
